refactor: Create semgrep.app.registry
