+++ semgrep/semgrep/commands/scan.py
@@ -18,7 +18,7 @@ from click_option_group import optgroup
 
 from semgrep import __VERSION__
 from semgrep import bytesize
-from semgrep.config_resolver import list_current_public_rulesets
+from semgrep.app.registry import list_current_public_rulesets
 from semgrep.constants import Colors
 from semgrep.constants import DEFAULT_MAX_CHARS_PER_LINE
 from semgrep.constants import DEFAULT_MAX_LINES_PER_FINDING
